init struct battery

+++ server/drivers/viacast_lcd_utils.h
@@ -1,4 +1,5 @@
 #include <dirent.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -13,9 +14,9 @@
 typedef struct {
     uint8_t battery_values[SIZE];
     uint8_t battery_current;
-    int head;  // Points to the position to insert the next element
+    uint8_t head;  // Points to the position to insert the next element
 } Battery;
 
 int filter(const struct dirent *name);
 
-void check_inotify_event(struct inotify_event *i, int *reload_icons);
+void check_inotify_event(struct inotify_event *i, int *reload_icons);
